fixed: date format in table and reorganization of the files
