+++ src/Components/Table/columns.js
@@ -1,13 +1,3 @@
-import format from "date-fns/format";
-
-//Format date to have a better UI
-const formatDate = ({ value }) => {
-  if (value === "0000/00/00") {
-    return "00/00/0000";
-  }
-  return format(new Date(value), "dd/MM/yyyy");
-};
-
 export const COLUMNS = [
   {
     Header: "First Name",
@@ -20,13 +10,9 @@ export const COLUMNS = [
   {
     Header: "Start Date",
     //Format date in order to sort it correctly and return something in case of empty field to avoid crash
-    accessor: (row) => {
-        if (row.startDate === undefined || null) {
-            return "0000/00/00";
-          }  
-      return format(new Date(row.startDate), "yyyy/MM/dd");
-    },
-    Cell: formatDate,
+    accessor: row => new Date(row.startDate),
+    sortType: 'datetime',
+    Cell: ({ cell: { value }}) => value.toLocaleDateString(),
   },
   {
     Header: "Department",
@@ -35,13 +21,9 @@ export const COLUMNS = [
   {
     Header: "Date of Birth",
     //Format date in order to sort it correctly and return something in case of empty field to avoid crash
-    accessor: (row) => {
-      if (row.birthDate === undefined || null) {
-        return "0000/00/00";
-      }
-      return format(new Date(row.birthDate), "yyyy/MM/dd");
-    },
-    Cell: formatDate,
+    accessor: row => new Date(row.birthDate),
+    sortType: 'datetime',
+    Cell: ({ cell: { value }}) => value.toLocaleDateString(),
   },
   {
     Header: "Street",
